fix type of several generated function parameters
These are all using an `int` parameter to pass something that is later used as a `size_t`.

This has potential to impact user code if that user code is using -Werror=sign-conversion and
passing non-literal `int` values to these generated functions.  I don't know whether any of these
generated functions would ever be called by a user (or have arguments passed verbatim via
macro), but even if so I suspect anyone that would run into this particular issue would welcome
the change, though, because they would have already had to go out of their way to suppress
the errors that this fixes.